Updates in extracting, and better error messages

+++ circus/extracting.py
@@ -156,14 +156,16 @@ def main(filename, params, nb_cpu, nb_gpu, use_gpu):
             offsets[i+1] = comm.bcast(numpy.array([local_nb_clusters], dtype=numpy.float32), root=i)
         node_pad   = numpy.sum(offsets[:comm.rank+1])        
         hfile      = h5py.File(file_out_suff + '.templates.hdf5', 'w', driver='mpio', comm=comm, libver='latest')
-        templates  = hfile.create_dataset('templates', shape=(N_e, N_t, 2*total_nb_clusters), dtype=numpy.float32, chunks=True)
+        templates  = hfile.create_dataset('templates', shape=(N_e, N_t, 2*total_nb_clusters), dtype=numpy.float32, chunks=(N_e, N_t, 1))
+        norms      = hfile.create_dataset('norms', shape=(2*total_nb_clusters, ), dtype=numpy.float32, chunks=True)
         electrodes = hfile.create_dataset('electrodes', shape=(total_nb_clusters, ), dtype=numpy.int32, chunks=True)
         amps_lims  = hfile.create_dataset('limits', shape=(total_nb_clusters, 2), dtype=numpy.float32, chunks=True)
     else:
         node_pad   = 0
         hfile      = h5py.File(file_out_suff + '.templates-%d.hdf5' %comm.rank, 'w', libver='latest')
-        templates  = hfile.create_dataset('templates', shape=(N_e, N_t, 2*local_nb_clusters), dtype=numpy.float32, chunks=True)
+        templates  = hfile.create_dataset('templates', shape=(N_e, N_t, 2*local_nb_clusters), dtype=numpy.float32, chunks=(N_e, N_t, 1))
         electrodes = hfile.create_dataset('electrodes', shape=(local_nb_clusters, ), dtype=numpy.int32, chunks=True)
+        norms      = hfile.create_dataset('norms', shape=(2*local_nb_clusters, ), dtype=numpy.float32, chunks=True)
         amps_lims  = hfile.create_dataset('limits', shape=(local_nb_clusters, 2), dtype=numpy.float32, chunks=True)
     
     cfile           = h5py.File(file_out_suff + '.clusters-%d.hdf5' %comm.rank, 'w', libver='latest')
@@ -178,17 +180,21 @@ def main(filename, params, nb_cpu, nb_gpu, use_gpu):
             data             = result['data_tmp_' + str(temp)].reshape(n_data, basis_proj.shape[1], N_e)
             first_component  = numpy.median(data, axis=0)
             tmp_templates    = numpy.dot(first_component.T, basis_rec)
-            tmpidx           = numpy.where(tmp_templates == tmp_templates.min())
-            temporal_shift   = template_shift - tmpidx[1][0]
-            electrodes      += [tmpidx[0][0]]
+            tmpidx           = divmod(tmp_templates.argmin(), tmp_templates.shape[1])
+            shift            = template_shift - tmpidx[1]
+            electrodes      += [indices[tmpidx[0][0]]]
             indices          = inv_nodes[edges[nodes[electrodes[-1]]]]
+            sorted_indices   = numpy.argsort(indices)
+            sindices         = indices[sorted_indices]
 
-            if temporal_shift > 0:
-                templates[indices, temporal_shift:, count_templates] = tmp_templates[indices, :-temporal_shift]
-            elif temporal_shift < 0:
-                templates[indices, :temporal_shift, count_templates] = tmp_templates[indices, -temporal_shift:]
+            if shift > 0:
+                templates[sindices, shift:, count_templates] = tmp_templates[sorted_indices, :-shift]
+            elif shift < 0:
+                templates[sindices, :shift, count_templates] = tmp_templates[sorted_indices, -shift:]
             else:
-                templates[indices, :, count_templates] = tmp_templates[indices, :]
+                templates[sindices, :, count_templates] = tmp_templates[sorted_indices]
+
+            norms[count_templates] = numpy.sqrt(numpy.mean(numpy.mean(templates[:,:,count_templates]**2,0),0))
 
             x, y, z          = data.shape
             data_flat        = data.reshape(x, y*z)
@@ -199,7 +205,7 @@ def main(filename, params, nb_cpu, nb_gpu, use_gpu):
                 data_flat[i, :] -= amplitudes[i]*first_flat[:, 0]
 
             variations       = 10*numpy.median(numpy.abs(amplitudes - numpy.median(amplitudes)))
-            physical_limit   = noise_thr*(-thresholds[tmpidx[0][0]])/tmp_templates.min()
+            physical_limit   = noise_thr*(-thresholds[indices[tmpidx[0][0]]])/tmp_templates.min()
             amp_min          = max(physical_limit, numpy.median(amplitudes) - variations)
             amp_max          = min(amp_limits[1], numpy.median(amplitudes) + variations)
             amps_lims[count_templates] = [amp_min, amp_max]
@@ -211,14 +217,16 @@ def main(filename, params, nb_cpu, nb_gpu, use_gpu):
             else:
                 second_component = data_flat.reshape(y, z)/numpy.sum(data_flat**2)
             
-            tmp_templates        = numpy.dot(second_component.T, basis_rec)
-            
-            if temporal_shift > 0:
-                templates[indices, temporal_shift:, templates.shape[2]/2 + count_templates] = tmp_templates[indices, :-temporal_shift]
-            elif temporal_shift < 0:
-                templates[indices, :temporal_shift, templates.shape[2]/2 + count_templates] = tmp_templates[indices, -temporal_shift:]
+            tmp_templates = numpy.dot(second_component.T, basis_rec)
+            offset        = templates.shape[2]/2 + count_templates
+            if shift > 0:
+                templates[sindices, shift:, offset] = tmp_templates[sorted_indices, :-shift]
+            elif shift < 0:
+                templates[sindices, :shift, offset] = tmp_templates[sorted_indices, -shift:]
             else:
-                templates[indices, :, templates.shape[2]/2 + count_templates] = tmp_templates[indices, :]
+                templates[sindices, :, offset] = tmp_templates[sorted_indices]
+
+            norms[offset] = numpy.sqrt(numpy.mean(numpy.mean(templates[:,:,offset]**2,0),0))
 
             count_templates += 1
 
@@ -256,8 +264,9 @@ def main(filename, params, nb_cpu, nb_gpu, use_gpu):
             n_clusters = numpy.sum([ts[i].get('templates').shape[2] for i in xrange(comm.size)])/2
             hfile      = h5py.File(file_out_suff + '.templates.hdf5', 'w', libver='latest')
             cfile      = h5py.File(file_out_suff + '.clusters.hdf5', 'w', libver='latest')
-            templates  = hfile.create_dataset('templates', shape=(N_e, N_t, 2*n_clusters), dtype=numpy.float32, chunks=True)
+            templates  = hfile.create_dataset('templates', shape=(N_e, N_t, 2*n_clusters), dtype=numpy.float32, chunks=(N_e, N_t, 1))
             electrodes = hfile.create_dataset('electrodes', shape=(n_clusters, ), dtype=numpy.int32, chunks=True)
+            norms      = hfile.create_dataset('norms', shape=(2*n_clusters, ), dtype=numpy.float32, chunks=True)
             amplitudes = hfile.create_dataset('limits', shape=(n_clusters, 2), dtype=numpy.float32, chunks=True)
             count      = 0
             for i in xrange(comm.size):
@@ -265,6 +274,8 @@ def main(filename, params, nb_cpu, nb_gpu, use_gpu):
                 middle      = loc_temp.shape[2]/2
                 templates[:,:,count:count+middle] = loc_temp[:,:,:middle]
                 templates[:,:,n_clusters+count:n_clusters+count+middle] = loc_temp[:,:,middle:]
+                norms[count:count+middle]                               = loc_norms[:middle]
+                norms[n_clusters+count:n_clusters+count+middle]         = loc_norms[middle:]
                 electrodes[count:count+middle] = ts[i].get('electrodes')
                 amplitudes[count:count+middle] = ts[i].get('limits')
                 count      += middle
@@ -286,10 +297,12 @@ def main(filename, params, nb_cpu, nb_gpu, use_gpu):
     merged1 = algo.merging_cc(comm, params, cc_merge, parallel_hdf5)
 
     comm.Barrier()
-    if comm.rank == 0:
-        print "Removing mixtures..."
-
-    merged2 = algo.delete_mixtures(comm, params, parallel_hdf5)
+    if remove_mixture:
+        if comm.rank == 0:
+            print "Removing mixtures..."
+        merged2 = algo.delete_mixtures(comm, params, parallel_hdf5)
+    else:
+        merged2 = [0, 0]
 
     if comm.rank == 0:
         io.print_info(["Number of global merges    : %d" %merged1[1],
